requirements.txt
aiohttp-proxy
async-timeout
backports.tarfile
better-proxy
crispy-bootstrap5
dask
dj-database-url
django-email
django-environ
django-filter
djangorestframework
environ
fqdn
imagehash
importlib-metadata
importlib-resources
ipympl
isoduration
jaraco.collections
jovian
jsonpointer
jupyter
lab
loguru
markdown
maturin
opencv-contrib-python
openpyxl
pandas-profiling
pip-chill
pipenv
plyer
pmdarima
psycopg2
psycopg2-binary
pyautogui
pydantic-settings
pyrogram
tgcrypto
tomli
unrar
uri-template
webcolors
